+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Charge Pads now accept four rechargeable items in their UI while continuing to charge carried and worn items when stood on. ([Issue 2123](https://github.com/FTBTeam/FTB-Mods-Issues/issues/2123))
 * Scrap Boxes now give a random reward when used. Reward selection is based on [PR 56](https://github.com/FTBTeam/FTB-Industrial-Contraptions/pull/56) by @jshipley. ([Issue 2124](https://github.com/FTBTeam/FTB-Mods-Issues/issues/2124))
 * The Fluid Cell guide now correctly explains that filled cells with identical contents can stack. ([Issue 2125](https://github.com/FTBTeam/FTB-Mods-Issues/issues/2125))
+* Charged Quantum Chestplates now enable hovering flight with a double tap of jump. Flight stops when the chestplate is removed or runs out of zaps. ([Issue 2128](https://github.com/FTBTeam/FTB-Mods-Issues/issues/2128))
 
 ### Added